move common and plugins out from app
