+++ setup.cfg
@@ -56,7 +56,7 @@ aiidalab_qe.app.static = *
 
 [options.entry_points]
 aiidalab_qe.properties =
-    bands = aiidalab_qe.app.plugins.bands:bands
+    bands = aiidalab_qe.plugins.bands:bands
 
 [aiidalab]
 title = Quantum ESPRESSO
