+++ src/aiidalab_qe/app/__init__.py
@@ -1,1 +1,7 @@
 """Package for the AiiDAlab QE app."""
+
+from .main import App
+
+__all__ = [
+    "App",
+]
