+++ src/aiidalab_qe/common/widgets.py
