+++ src/aiidalab_qe/plugins/bands/result.py
@@ -3,7 +3,7 @@
 """
 
 
-from aiidalab_qe.app.common.panel import ResultPanel
+from aiidalab_qe.common.panel import ResultPanel
 
 
 def export_bands_data(outputs, fermi_energy=None):
