+++ src/aiidalab_qe/plugins/bands/setting.py
@@ -3,7 +3,7 @@
 import ipywidgets as ipw
 from aiida.orm import Int, Str
 
-from aiidalab_qe.app.common.panel import Panel
+from aiidalab_qe.common.panel import Panel
 
 
 class Setting(Panel):
